requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-oauth-toolkit==1.7.1
PyJWT==2.8.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
cryptography==41.0.7
django-extensions==3.2.3
factory-boy==3.3.0
pytest-django==4.7.0
pytest-cov==4.1.0
django-filter==23.3
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0